Move pages into pages/ dir.
